Make Manager fully templated
