+++ src/CMakeLists.txt
@@ -5,7 +5,6 @@ if(${NATIVE_ENVIRONMENT} STREQUAL "X11")
 endif()
 
 add_subdirectory(events)
-add_subdirectory(window)
 add_subdirectory(environment)
 
 add_executable(${PROJECT_NAME} main.cpp)
@@ -16,13 +15,11 @@ target_include_directories(${PROJECT_NAME}
 add_dependencies(${PROJECT_NAME}
 	events
 	environment
-	window
 )
 target_link_libraries(${PROJECT_NAME}
 	PRIVATE
 	events
 	environment
-	window
 )
 target_compile_options(${PROJECT_NAME}
 	PRIVATE
