--- src/window/CMakeLists.txt
@@ -1,6 +1,0 @@
-add_library(window OBJECT
-	${CMAKE_CURRENT_SOURCE_DIR}/Manager.cpp
-)
-target_include_directories(window PRIVATE
-	${TARGET_INCLUDE_DIRECTORY}
-)
